pyproject.toml
[tool.poetry]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = ["Seu nome aqui <seu.email@email.com.br>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "apps" },
    { include = "realmate_challenge" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = ">=5.1.6,<6.0.0"
celery = ">=5.3.0"
redis = ">=5.0.0"
django-celery-results = ">=2.5.0"
djangorestframework = ">=3.14.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
